Remove hardcoded filters for benefit cards
Before the production of zelda and mPOS for "MEIs" we do not share too much data from benefit cards and we can't find a solution (regex) that could allow us to change and filter only sodexo/vr/va/alelo cards. Actually, they buy different BIN ranges and we could only use the BIN as regex but it would take time as they have plenty of different ranges and are continuously changing and adding more (that would set for us continuous tasks of updating the list and regexes), and that is the case of Alelo mainly, the others we know very few about the patterns.